setup/requirements.txt
# фиксируем версии для предсказуемости
psycopg2-binary==2.9.9
